move hubconf models
